ec: Fix 32 bits issues on file size calculation
Some additional 32 bits issues have been added by a recent patch.
This patch solves it.

BUG: 1146903
Reviewed-on: http://review.gluster.org/8882
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: Emmanuel Dreyfus <manu@netbsd.org>

+++ xlators/cluster/ec/src/ec-data.h
@@ -149,8 +149,8 @@ struct _ec_lock
     int32_t              refs;
     int32_t              acquired;
     int32_t              have_size;
-    size_t               size;
-    size_t               size_delta;
+    uint64_t             size;
+    uint64_t             size_delta;
     uint64_t             version;
     uint64_t             version_delta;
     ec_fop_data_t *      owner;
